Update dependencies to LTS 14.5
- Bump version from 0.0.3 to 0.1.0.
- Update dependency on 'aeson' from <1.3 to <1.5.
- Update dependency on 'th-abstraction from <0.3 to <0.4.
- Update dependency on 'lens' from <4.17 to <4.18.
- Remove dependency on 'base-compat'.

As it's a little unclear why 'base-compat' was necessary, it's estimated
that it no longer is, since it compiles without and we are in the future
now.

The PVP states:

> Note that modifying imports or depending on a newer version of another
> package may cause extra non-orphan instances to be exported and thus
> force a minor version change.

As I am unsure if any of the bumped dependencies cause extra non-orphan
instances, but all of them bump the minor version, a safe choice is to
bump this package's minor version.